23-12-30
1.更新md文档
2.新增zbuffer消隐算法
3.新增光照
4.新增背面剔除

+++ 完整项目代码/Renderer(vc)/my_math/my_math.cpp
@@ -58,14 +58,14 @@ namespace XX_XZH {
     w_ = w;
   }
   void Vector3::Normalize() {
-    float tmp_result;
+    double tmp_result;
     tmp_result = this->Norm();
     this->x_ = this->x_ / tmp_result;
     this->y_ = this->y_ / tmp_result;
     this->z_ = this->z_ / tmp_result;
   }
-  float Vector3::Norm() {
-    return (float)sqrt(x_ * x_ + y_ * y_ + z_ * z_);
+  double Vector3::Norm() {
+    return sqrt(x_ * x_ + y_ * y_ + z_ * z_);
   }
   Vector3& Vector3::Identity()
   {
